requirements.txt
bcrypt==3.2.0
bidict==0.21.2
cffi==1.14.5
click==8.0.1
Flask==2.0.1
Flask-Login==0.5.0
Flask-SocketIO==5.1.0
Flask-SQLAlchemy==2.5.1
Flask-WTF==0.15.1
greenlet==1.1.0
itsdangerous==2.0.1
Jinja2==3.0.1
MarkupSafe==2.0.1
pbkdf2==1.3
py-bcrypt==0.4
pycparser==2.20
python-dotenv==0.17.1
python-engineio==4.2.0
python-socketio==5.3.0
six==1.16.0
SQLAlchemy==1.4.17
Werkzeug==2.0.1
WTForms==2.3.3
